requirements.txt
certifi==2025.6.15
cftime==1.6.4.post1
contourpy==1.3.2
cycler==0.12.1
DateTime==5.5
fonttools==4.58.4
kiwisolver==1.4.8
matplotlib==3.10.3
netCDF4==1.7.2
numpy==2.3.1
packaging==25.0
pillow==11.3.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-sofa==0.2.0
pytz==2025.2
scipy==1.16.0
setuptools==80.9.0
six==1.17.0
zope.interface==7.2
